+++ src/main/java/racingcar/Application.java
@@ -1,11 +1,11 @@
 package racingcar;
 
-import racingcar.util.RacingGame;
+import racingcar.handler.GameHandler;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        RacingGame racingGame = new RacingGame();
+        GameHandler racingGame = new GameHandler();
 
         racingGame.run();
     }
